+++ .idea/misc.xml
@@ -3,9 +3,13 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="..\:/Users/IMustQuestion/Documents/301workspace/QrACutie/app/src/main/res/layout/activity_main.xml" value="0.22" />
-        <entry key="..\:/Users/IMustQuestion/Documents/301workspace/QrACutie/app/src/main/res/layout/activity_player_collection.xml" value="0.33" />
-        <entry key="..\:/Users/IMustQuestion/Documents/301workspace/QrACutie/app/src/main/res/layout/list_qr_code.xml" value="0.5" />
+        <entry key="..\:/Users/Owner/OneDrive/Documents/Degree n stuff/Winter 2022/Cmput 301/QrACutie/app/src/main/res/layout/activity_camera.xml" value="0.24895833333333334" />
+        <entry key="..\:/Users/Owner/OneDrive/Documents/Degree n stuff/Winter 2022/Cmput 301/QrACutie/app/src/main/res/layout/activity_main.xml" value="0.24895833333333334" />
+        <entry key="..\:/Users/Owner/OneDrive/Documents/Degree n stuff/Winter 2022/Cmput 301/QrACutie/app/src/main/res/layout/after_scanning_qr_code.xml" value="0.2" />
+        <entry key="..\:/Users/hetul/Downloads/School/CMPUT301/Project/QrACutie/app/src/main/res/layout/activity_camera.xml" value="0.25" />
+        <entry key="..\:/Users/hetul/Downloads/School/CMPUT301/Project/QrACutie/app/src/main/res/layout/activity_main.xml" value="0.67" />
+        <entry key="..\:/Users/hetul/Downloads/School/CMPUT301/Project/QrACutie/app/src/main/res/layout/activity_maps.xml" value="0.22" />
+        <entry key="..\:/Users/hetul/Downloads/School/CMPUT301/Project/QrACutie/app/src/main/res/layout/activity_save_qr.xml" value="0.33" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.279922279872806" />
       </map>
     </option>
