+++ app/src/main/java/com/example/qracutie/SaveQR.java
@@ -23,18 +23,33 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.CheckBox;
 
+import com.google.firebase.firestore.FirebaseFirestore;
+
+import java.security.MessageDigest;
 import java.util.function.Consumer;
 
 public class SaveQR extends AppCompatActivity {
+
+    private FirebaseFirestore db = FirebaseFirestore.getInstance();
     static final int REQUEST_IMAGE_CAPTURE = 1;
     private static final int REQUEST_LOCATION_PERMISSION = 1;
     ActivityResultLauncher<Intent> activityResultLauncher;
     LocationManager locationManager;
+    GameQRCode scannedQrCode;
+    double scannedLatitude;
+    double scannedLongitude;
+    Boolean boxChecked = false;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_save_qr);
+
+        Intent intent = getIntent();
+        String qrCodeString = intent.getStringExtra("qrcode");
+        String qrCodeHash = shaHash(qrCodeString);
+        int points = computeHashScore(qrCodeHash);
+        scannedQrCode = new GameQRCode(qrCodeHash, points);
 
         locationManager = (LocationManager) getSystemService(LOCATION_SERVICE);
         // This part asks for location access permission from the user
@@ -54,6 +69,17 @@ public class SaveQR extends AppCompatActivity {
             }
         });
 
+        Button done = findViewById(R.id.done_button);
+        done.setOnClickListener(view -> {
+            // Main
+            if(boxChecked){
+                scannedQrCode.setLatitude(scannedLatitude);
+                scannedQrCode.setLongitude(scannedLongitude);
+            }
+            db.collection("GameQRCodes").document(scannedQrCode.getHash()).set(scannedQrCode);
+            Intent intentMain = new Intent(this, MainActivity.class);
+            startActivity(intentMain);
+        });
 
     }
 
@@ -67,15 +93,11 @@ public class SaveQR extends AppCompatActivity {
         switch (view.getId()) {
             case R.id.location:
                 if (checked) {
-                    if (ActivityCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(this, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
-                        // TODO: Consider calling
-                        //    ActivityCompat#requestPermissions
-                        // here to request the missing permissions, and then overriding
-                        //   public void onRequestPermissionsResult(int requestCode, String[] permissions,
-                        //                                          int[] grantResults)
-                        // to handle the case where the user grants the permission. See the documentation
-                        // for ActivityCompat#requestPermissions for more details.
-                        return;
+                    // This part asks for location access permission from the user
+                    if (ActivityCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
+                        ActivityCompat.requestPermissions(this, new String[]
+                                        {Manifest.permission.ACCESS_FINE_LOCATION},
+                                REQUEST_LOCATION_PERMISSION);
                     }
                     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
                         locationManager.getCurrentLocation(
@@ -85,7 +107,9 @@ public class SaveQR extends AppCompatActivity {
                                 new Consumer<Location>() {
                                     @Override
                                     public void accept(Location location) {
-                                        // code
+                                        scannedLatitude = location.getLatitude();
+                                        scannedLongitude = location.getLongitude();
+                                        boxChecked = true;
                                     }
                                 });
                     }
@@ -93,9 +117,70 @@ public class SaveQR extends AppCompatActivity {
                 // Put some meat on the sandwich
                 else {
                     // Remove the meat
+                    boxChecked = false;
                     break;
                 }
         }
+    }
 
+    // https://stackoverflow.com/questions/5531455/how-to-hash-some-string-with-sha256-in-java
+    public static String shaHash(final String base) {
+        try{
+            final MessageDigest digest = MessageDigest.getInstance("SHA-256");
+            final byte[] hash = digest.digest(base.getBytes("UTF-8"));
+            final StringBuilder hexString = new StringBuilder();
+            for (int i = 0; i < hash.length; i++) {
+                final String hex = Integer.toHexString(0xff & hash[i]);
+                if(hex.length() == 1)
+                    hexString.append('0');
+                hexString.append(hex);
+            }
+            return hexString.toString();
+        } catch(Exception ex){
+            throw new RuntimeException(ex);
+        }
+    }
+
+    private int charToInt(char ch)
+    {
+        if (ch >= '0' && ch <= '9')
+            return ch - '0';
+        if (ch >= 'A' && ch <= 'F')
+            return ch - 'A' + 10;
+        if (ch >= 'a' && ch <= 'f')
+            return ch - 'a' + 10;
+        return -1;
+    }
+
+    private int computeHashScore(String hash) {
+        int points = 0;
+        int indexIncrement = 0;
+        char currChar;
+        char nextChar;
+        int repeatCounter = 0;
+        for(int i = 0; i < hash.length();i++){
+            currChar = hash.charAt(i);
+            Boolean check = true;
+            indexIncrement = i+1;
+            while(check) {
+                nextChar = '/';
+                if(indexIncrement < hash.length()){
+                    nextChar = hash.charAt(indexIncrement);
+                }
+                if(nextChar == currChar){
+                    indexIncrement++;
+                    repeatCounter++;
+                    continue;
+                }
+                if(repeatCounter > 0){
+                    int baseValue = charToInt(currChar);
+                    points += (int) Math.pow(baseValue, repeatCounter);
+                    repeatCounter = 0;
+                    i = indexIncrement;
+                }
+                check = false;
+            }
+        }
+        return points;
     }
 }
