+++ app/src/main/res/layout/activity_camera.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
@@ -21,15 +20,6 @@
 
     </androidx.camera.view.PreviewView>
 
-    <Button
-        android:id="@+id/qrCodeFoundButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="QR Code Found"
-        app:layout_constraintBottom_toBottomOf="@+id/PeekView"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
     <RelativeLayout
         android:id="@+id/relativeLayout"
         android:layout_width="match_parent"
@@ -40,17 +30,19 @@
             android:layout_height="300dp"
             android:layout_centerInParent="true"
             android:background="@drawable/border" />
+
         <Button
-            android:id="@+id/button"
+            android:id="@+id/qrCodeFoundButton"
             android:layout_width="wrap_content"
             android:layout_height="85dp"
             android:layout_alignParentEnd="true"
             android:layout_alignParentBottom="true"
             android:layout_marginEnd="157dp"
             android:layout_marginBottom="41dp"
-            android:background="@drawable/ic_baseline_filter_tilt_shift_24"
-            android:backgroundTint="#D1CECE" />
-    </RelativeLayout>
+            android:text="QR Code Found"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent" />
+</RelativeLayout>
 
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
