+++ app/src/main/res/layout/activity_save_qr.xml
@@ -32,12 +32,13 @@
         android:layout_gravity="center"
         android:text="Capture">
     </Button>
+
     <Button
+        android:id="@+id/done_button"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_gravity="center"
-        android:text="Done">
-    </Button>
+        android:text="Done"></Button>
 
 
 </LinearLayout>
